polish

+++ proj2/part4.c
@@ -9,10 +9,16 @@
 
 
 volatile sig_atomic_t alarm_fired = 0;
+volatile sig_atomic_t child_changed = 0;
 
 void handle_alarm(int sig) {
   (void)sig;
   alarm_fired = 1;
+}
+
+void handle_sigchld(int sig) {
+  (void)sig;
+  child_changed = 1;
 }
 
 int find_next_runnable(ProcessList* process_list, int current) {
@@ -301,6 +307,15 @@ int main(int argc, char* argv[]) {
           exit(EXIT_FAILURE);
         }
         printf("[child %d] received SIGUSR1, execing %s\n", getpid(), command_lines.list[i]);
+        if (sigprocmask(SIG_SETMASK, &oldset, NULL) == -1) {
+          perror("sigprocmask restore child");
+          free(line);
+          free_process_list(&process_list);
+          free_str_list(&command_lines);
+          free_str_list(&args);
+          fclose(in);
+          _exit(EXIT_FAILURE);
+        }
         execvp(args.list[0], args.list);
         perror("execvp");
         free(line);
@@ -339,16 +354,13 @@ int main(int argc, char* argv[]) {
     return 0;
   }
   int current = 0;
-  if (kill(process_list.processes[current].pid, SIGUSR1) == -1) {
-    perror("kill SIGUSR1");
-  }
-  process_list.processes[current].status = RUNNING;
 
   sigset_t blockset;
 
   sigemptyset(&blockset);
   sigemptyset(&oldset);
   sigaddset(&blockset, SIGALRM);
+  sigaddset(&blockset, SIGCHLD);
 
   if (sigprocmask(SIG_BLOCK, &blockset, &oldset) == -1) {
     perror("sigprocmask block");
@@ -357,10 +369,21 @@ int main(int argc, char* argv[]) {
     fclose(in);
     return 1;
   }
-  struct sigaction sa = {0};
-  sa.sa_handler = handle_alarm;
-  if (sigaction(SIGALRM, &sa, NULL) == -1) {
-    perror("sigaction");
+  struct sigaction sa_alarm = {0};
+  sa_alarm.sa_handler = handle_alarm;
+  if (sigaction(SIGALRM, &sa_alarm, NULL) == -1) {
+    perror("sigaction SIGALRM");
+    free(line);
+    free_process_list(&process_list);
+    fclose(in);
+    return 1;
+  }
+
+  struct sigaction sa_chld = {0};
+  sa_chld.sa_handler = handle_sigchld;
+  sa_chld.sa_flags = SA_NOCLDSTOP;
+  if (sigaction(SIGCHLD, &sa_chld, NULL) == -1) {
+    perror("sigaction SIGCHLD");
     free(line);
     free_process_list(&process_list);
     fclose(in);
@@ -369,45 +392,78 @@ int main(int argc, char* argv[]) {
 
   sigset_t suspend_mask = oldset;
   sigdelset(&suspend_mask, SIGALRM);
+  sigdelset(&suspend_mask, SIGCHLD);
+
+  printf("[MCP] starting PID %d\n", process_list.processes[current].pid);
+  if (kill(process_list.processes[current].pid, SIGUSR1) == -1) {
+    perror("kill SIGUSR1");
+  }
+  else {
+    process_list.processes[current].status = RUNNING;
+  }
 
   alarm(2);
   while (process_list.active > 0) {
-    while (!alarm_fired) {
+    while (!alarm_fired && !child_changed) {
       if (sigsuspend(&suspend_mask) == -1 && errno != EINTR) {
         perror("sigsuspend");
         exit(EXIT_FAILURE);
       }
     }
+
+    int time_slice_expired = alarm_fired;
+    int child_state_changed = child_changed;
     alarm_fired = 0;
-    printf("[MCP] alarm fired, time slice expired for PID %d\n", process_list.processes[current].pid);
-    reap_exited(&process_list);
-    // need to check if current process is still running
-    // int exited = is_process_exited(process_list.processes[current].pid);
-    // if (exited == -1) {
-    //   perror("is_process_exited");
-    // }
-    if (process_list.processes[current].status != EXITED) {
-      printf("[MCP] stopping PID %d\n", process_list.processes[current].pid);
-      if (kill(process_list.processes[current].pid, SIGSTOP) == -1) {
-        perror("kill SIGSTOP");
+    child_changed = 0;
+
+    if (child_state_changed) {
+      reap_exited(&process_list);
+    }
+
+    if (process_list.active <= 0) {
+      break;
+    }
+
+    if (!time_slice_expired && process_list.processes[current].status != EXITED) {
+      continue;
+    }
+
+    if (time_slice_expired) {
+      printf("[MCP] alarm fired, time slice expired for PID %d\n", process_list.processes[current].pid);
+      reap_exited(&process_list);
+
+      if (process_list.processes[current].status != EXITED) {
+        printf("[MCP] stopping PID %d\n", process_list.processes[current].pid);
+        if (kill(process_list.processes[current].pid, SIGSTOP) == -1) {
+          if (errno == ESRCH) {
+            reap_exited(&process_list);
+          }
+          else {
+            perror("kill SIGSTOP");
+          }
+        }
+        else {
+          process_list.processes[current].status = STOPPED;
+        }
       }
-      else {
-        process_list.processes[current].status = STOPPED;
-      }
+    }
+    else {
+      alarm(0);
+      printf("[MCP] PID %d exited before its time slice expired\n", process_list.processes[current].pid);
     }
 
     print_status(&process_list);
 
-    reap_exited(&process_list);
-    // find next process
+    if (process_list.active <= 0) {
+      break;
+    }
+
     current = find_next_runnable(&process_list, current);
     if (current == -1) {
       continue;
     }
     printf("[MCP] next runnable process is PID %d\n", process_list.processes[current].pid);
-    // based on the current state, determine signal to send
-    // NEVER_RUN -> SIGUSR1
-    //
+
     if (process_list.processes[current].status == NEVER_RUN) {
       printf("[MCP] starting PID %d\n", process_list.processes[current].pid);
       if (kill(process_list.processes[current].pid, SIGUSR1) == -1) {
@@ -428,6 +484,7 @@ int main(int argc, char* argv[]) {
     }
     alarm(2);
   }
+  alarm(0);
   printf("[MCP] all processes exited, MCP exiting\n");
   reap_exited(&process_list);
